Rust 1.2.0-dev
19a39710c

Rust Documentation

Welcome to the Rust documentation! You can use the section headings above to jump to any particular section.

1 Getting Started

If you haven't seen Rust at all yet, the first thing you should read is the introduction to The Rust Programming Language. It'll give you a good idea of what Rust is like.

The book provides a lengthy explanation of Rust, its syntax, and its concepts. Upon completing the book, you'll be an intermediate Rust developer, and will have a good grasp of the fundamental ideas behind Rust.

Rust By Example was originally a community resource, but was then donated to the Rust project. As the name implies, it teaches you Rust through a series of small examples.

2 Community & Getting Help

If you need help with something, or just want to talk about Rust with others, there are a few places you can do that:

The Rust IRC channels on irc.mozilla.org are the fastest way to get help. #rust is the general discussion channel, and you'll find people willing to help you with any questions you may have.

There are also three specialty channels: #rust-gamedev and #rust-osdev are for game development and operating system development, respectively. There's also #rust-internals, which is for discussion of the development of Rust itself.

You can also get help on Stack Overflow. Searching for your problem might reveal someone who has asked it before!

There is an active subreddit with lots of discussion and news about Rust.

There is also a user forum, for all user-oriented discussion, and a developer forum, where the development of Rust itself is discussed.

3 Specification

Rust does not have an exact specification, but an effort to describe as much of the language in as much detail as possible is in the reference.

4 Tools

Rust is still a young language, so there isn't a ton of tooling yet, but the tools we have are really nice.

Cargo is Rust's package manager, and its website contains lots of good documentation.

rustdoc is used to generate documentation for Rust code.

5 FAQs

There are questions that are asked quite often, so we've made FAQs for them:

Language Design FAQ

Language FAQ

Project FAQ

How to submit a bug report

6 The Standard Library

We have API documentation for the entire standard library. There's a list of crates on the left with more specific sections, or you can use the search bar at the top to search for something if you know its name.

7 The Error Index

If you encounter an error while compiling your code you may be able to look it up in the Rust Compiler Error Index.

8 Community Translations

Several projects have been started to translate the documentation into other languages:

Russian

Korean

Chinese

Spanish