Rust 1.12.0-dev
c267ab4e3

Rust Documentation

This is an index of the documentation included with the Rust compiler. For more comprehensive documentation see the website.

The Rust Programming Language. Also known as "The Book", The Rust Programming Language is the most comprehensive resource for all topics related to Rust, and is the primary official document of the language.

The Rust Reference. While Rust does not have a specification, the reference tries to describe its working in detail. It tends to be out of date.

Standard Library API Reference. Documentation for the standard library.

The Rustonomicon. An entire book dedicated to explaining how to write unsafe Rust code. It is for advanced Rust programmers.

Compiler Error Index. Extended explanations of the errors produced by the Rust compiler.